dashboard/requirements.txt
google-cloud-bigquery>=3.40.1
pandas>=2.2.3
plotly>=6.0.1
streamlit>=1.55.0
